Refactor CreateModal to use PostEditor and confirmation logic
