+++ src/components/PostEditorModal/index.js
@@ -1,3 +1,3 @@
-import PostEditorModal from './PostEditorModal'
-import connector from './PostEditorModal.connector'
-export default connector(PostEditorModal)
+import component from './PostEditorModal'
+
+export default component
